+++ src/main/java/miyucomics/hexpose/actions/display/OpCompareStyles.kt
@@ -8,5 +8,5 @@ import miyucomics.hexpose.iotas.display.getDisplay
 
 object OpCompareStyles : ConstMediaAction {
 	override val argc = 2
-	override fun execute(args: List<Iota>, env: CastingEnvironment) = (args.getDisplay(0, argc).text.style == args.getDisplay(1, argc).text.style).asActionResult
+	override fun execute(args: List<Iota>, env: CastingEnvironment) = (args.getDisplay(0, argc).style == args.getDisplay(1, argc).style).asActionResult
 }